Pointer math fixes.
git-svn-id: https://svn.mcs.anl.gov/repos/parallel-netcdf/trunk@56 c368ceb8-fd28-0410-a457-e46ee1d7cd4a

+++ src/lib/validator.c
@@ -146,7 +146,7 @@ val_get_NC_string(bufferinfo *gbp, NC_string **ncstrpp) {
   } 
   if (memcmp(gbp->pos, pad, padding) != 0) {
     printf("Error @ [0x%8.8Lx]: \n\tPadding should be 0x00 for the name string alignment of ",
-	   (size_t)(gbp->pos - gbp->base) + gbp->offset - gbp->size);
+	   ((size_t) gbp->pos - (size_t) gbp->base) + gbp->offset - gbp->size);
     free_NC_string(ncstrp);
     return EINVAL;
   }
@@ -209,14 +209,14 @@ val_get_NC_dimarray(bufferinfo *gbp, NC_dimarray *ncap) {
   if(ncap->nelems == 0) {
     if (type != NC_DIMENSION && type != NC_UNSPECIFIED) {
       printf("Error @ [0x%8.8Lx]: \n\tInvalid NC component type, while ",
-	      (size_t)(gbp->pos - gbp->base) + gbp->offset - gbp->size - 2 * X_SIZEOF_SIZE_T);
+	      ((size_t) gbp->pos - (size_t) gbp->base) + gbp->offset - gbp->size - 2 * X_SIZEOF_SIZE_T);
       printf("NC_DIMENSION or NC_UNSPECIFIED is expected for ");
       return EINVAL;
     }
   } else {
     if(type != NC_DIMENSION) {
       printf("Error @ [0x%8.8Lx]: \n\tInvalid NC component type, while ",
-	      (size_t)(gbp->pos - gbp->base) + gbp->offset - gbp->size - 2 * X_SIZEOF_SIZE_T);
+	      ((size_t) gbp->pos - (size_t) gbp->base) + gbp->offset - gbp->size - 2 * X_SIZEOF_SIZE_T);
       printf("NC_DIMENSION is expected since number of dimensions is %d for ", ncap->nelems);
       return EINVAL;
     }
@@ -263,7 +263,7 @@ val_get_nc_type(bufferinfo *gbp, nc_type *typep) {
       && type != NC_FLOAT
       && type != NC_DOUBLE) {
     printf("Error @ [0x%8.8Lx]: \n\tUnknown data type for the values of ",
-	   (size_t)(gbp->pos - gbp->base) + gbp->offset - gbp->size - X_SIZEOF_INT);
+	   ((size_t) gbp->pos - (size_t) gbp->base) + gbp->offset - gbp->size - X_SIZEOF_INT);
     return EINVAL; 
   }
  
@@ -307,7 +307,7 @@ val_get_NC_attrV(bufferinfo *gbp, NC_attr *attrp) {
   memset(pad, 0, X_ALIGN-1);
   if (memcmp(gbp->pos, pad, padding) != 0) {
     printf("Error @ [0x%8.8Lx]: \n\tPadding should be 0x00 for the values alignment of ",
-           (size_t)(gbp->pos - gbp->base) + gbp->offset - gbp->size); 
+           ((size_t) gbp->pos - (size_t) gbp->base) + gbp->offset - gbp->size); 
     return EINVAL;
   }
   gbp->pos = (void *)((char *)gbp->pos + padding);
@@ -385,14 +385,14 @@ val_get_NC_attrarray(bufferinfo *gbp, NC_attrarray *ncap){
   if(ncap->nelems == 0) {
     if (type != NC_ATTRIBUTE && type != NC_UNSPECIFIED) {
       printf("Error @ [0x%8.8Lx]: \n\tInvalid NC component type, while ",
-              (size_t)(gbp->pos - gbp->base) + gbp->offset - gbp->size - 2 * X_SIZEOF_SIZE_T);
+              ((size_t) gbp->pos - (size_t) gbp->base) + gbp->offset - gbp->size - 2 * X_SIZEOF_SIZE_T);
       printf("NC_ATTRIBUTE or NC_UNSPECIFIED is expected for "); 
       return EINVAL;
     }
   } else {
     if(type != NC_ATTRIBUTE) {
       printf("Error @ [0x%8.8Lx]: \n\tInvalid NC component type, while ",
-              (size_t)(gbp->pos - gbp->base) + gbp->offset - gbp->size - 2 * X_SIZEOF_SIZE_T);
+              ((size_t) gbp->pos - (size_t) gbp->base) + gbp->offset - gbp->size - 2 * X_SIZEOF_SIZE_T);
       printf("NC_ATTRIBUTE is expected since number of attributes is %d for ", ncap->nelems);  
       return EINVAL;
     }
@@ -521,14 +521,14 @@ val_get_NC_vararray(bufferinfo *gbp, NC_vararray *ncap) {
   if(ncap->nelems == 0) {
     if (type != NC_VARIABLE && type != NC_UNSPECIFIED) {
       printf("Error @ [0x%8.8Lx]: \n\tInvalid NC component type, while ",
-              (size_t)(gbp->pos - gbp->base) + gbp->offset - gbp->size - 2 * X_SIZEOF_SIZE_T);
+              ((size_t) gbp->pos - (size_t) gbp->base) + gbp->offset - gbp->size - 2 * X_SIZEOF_SIZE_T);
       printf("NC_VARIABLE or NC_UNSPECIFIED is expected for ");
       return EINVAL;
     }
   } else {
     if(type != NC_VARIABLE) {
       printf("Error @ [0x%8.8Lx]: \n\tInvalid NC component type, while ",
-              (size_t)(gbp->pos - gbp->base) + gbp->offset - gbp->size - 2 * X_SIZEOF_SIZE_T);
+              ((size_t) gbp->pos - (size_t) gbp->base) + gbp->offset - gbp->size - 2 * X_SIZEOF_SIZE_T);
       printf("NC_VARIABLE is expected since number of variables is %d for ", ncap->nelems);        
       return EINVAL;
     }
